+++ fishyrl/__init__.py
@@ -1,5 +1,9 @@
-from . import actions, buffers, distributions, losses, models, utilities
-
-
-# https://github.com/Eclectic-Sheep/sheeprl/tree/main
-# https://duarteocarmo.com/blog/opinionated-python-boilerplate
+from . import (
+    actions,
+    buffers,
+    distributions,
+    dreamer,
+    losses,
+    models,
+    utilities,
+)